requirements.txt
anthropic==0.93.0
python-dotenv==1.2.2
python-telegram-bot==22.7
